refactor(domains): move favorites helper to domains/catalog (#232)
