+++ src/domains/catalog/favorites.ts
